+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## master (unreleased)
 
+### New Feature
+
+* Add more details in `WebDriver#open` error
+
+### Changes
+
+* Remove extra refresh if `WebDriver#open` failed
+
 ## 0.16.1 (2021-07-26)
 
 ### Fixes